03-Azure/01-04-AI/01_Azure_OpenAI_Service/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

cryptography
azure-functions
azure-storage-blob
azure-identity
requests
azure-identity
azure-keyvault-secrets
azure-ai-formrecognizer
openai
num2words
matplotlib 
plotly 
scipy 
scikit-learn